Fixed segmenting logic
Previously, the segments were consistent but the data was not being segmented correctly, for example the midi and mxl segments would both be 14 but the first 7 segments would contain all the music while the last 7 would not. Segmenting logic was flipped to use the midi tempo changes instead of the musicxml measure boundaries for determining where the split points are. The changes also allow the segmenter to keep the clean segments from corrupt files to maximise the amount of data available.